apps/post/serializers/post_category.py
from rest_framework import serializers

from ..models import PostCategory


class PostCategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = PostCategory
        fields = "__all__"
