requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0.3
django-ckeditor==6.7.1
django-filter==23.5
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
whitenoise==6.6.0
